Chore: Restructuring the files
