+++ requirements.txt
@@ -4,4 +4,3 @@ python-dotenv==1.0.1
 gunicorn==21.2.0
 python-dotenv==1.0.1
 pydantic==1.10.12
-telex
